fix: project selector now assigns to project
feat: project selector can now handle no project input

commit 20250620 @ 1601
