+++ cmd/issues_modify.go
@@ -1,0 +1,30 @@
+package cmd
+
+import (
+	"fmt"
+	"os"
+
+	"github.com/spf13/cobra"
+)
+
+// modifyCmd represents the modify command
+var modifyCmd = &cobra.Command{
+	Use:   "modify [issue-id]",
+	Short: "Modify an existing Linear issue",
+	Long:  `Modifies an existing Linear issue identified by its ID.`,
+	Args:  cobra.ExactArgs(1), // Requires exactly one argument (the issue ID)
+	Run: func(cmd *cobra.Command, args []string) {
+		issueID := args[0]
+		fmt.Printf("Modify issue called for ID: %s\n", issueID)
+		fmt.Println("TODO: Implement issue modification logic")
+		// Exit with a non-zero status code to indicate unimplemented state
+		os.Exit(1)
+	},
+}
+
+// init function is not strictly necessary if you register commands in issues.go's init
+/*
+func init() {
+	// No specific flags defined for the modify command in the original code
+}
+*/
